now im able to write ui tests -- base logic in presenters added
