+++ app/src/main/java/ru/popov/bodya/howmoney/presentation/mvp/about/AboutView.kt
@@ -6,7 +6,5 @@ import ru.popov.bodya.core.mvp.AppView
  *  @author popovbodya
  */
 interface AboutView : AppView {
-
     fun showAppVersion(version: String)
-
 }
