+++ app/src/main/java/ru/popov/bodya/howmoney/presentation/mvp/account/AccountView.kt
@@ -6,6 +6,4 @@ import ru.popov.bodya.core.mvp.AppView
  *  @author popovbodya
  */
 interface AccountView : AppView {
-
-
 }
