--- app/src/main/java/ru/popov/bodya/howmoney/presentation/mvp/enrollment/EnrollmentView.kt
@@ -1,14 +1,0 @@
-package ru.popov.bodya.howmoney.presentation.mvp.enrollment
-
-import ru.popov.bodya.core.mvp.AppView
-import ru.popov.bodya.howmoney.domain.enrollment.models.EnrollmentCategory
-import ru.popov.bodya.howmoney.domain.enrollment.models.EnrollmentCategoryBalance
-
-/**
- *  @author popovbodya
- */
-interface EnrollmentView: AppView {
-
-    fun showDiagram(enrollmentCategoryMap: Map<EnrollmentCategory, Double>, finalBalance: Double)
-
-}
