--- app/src/main/java/ru/popov/bodya/howmoney/presentation/mvp/expense/ExpenseView.kt
@@ -1,13 +1,0 @@
-package ru.popov.bodya.howmoney.presentation.mvp.expense
-
-import ru.popov.bodya.core.mvp.AppView
-import ru.popov.bodya.howmoney.domain.expense.models.ExpenseCategory
-
-/**
- *  @author popovbodya
- */
-interface ExpenseView: AppView {
-
-    fun showDiagram(expenseCategoryMap: Map<ExpenseCategory, Double>, finalBalance: Double)
-
-}
