--- app/src/main/java/ru/popov/bodya/howmoney/presentation/mvp/replenishment/ReplenishmentView.kt
@@ -1,11 +1,0 @@
-package ru.popov.bodya.howmoney.presentation.mvp.replenishment
-
-import ru.popov.bodya.core.mvp.AppView
-
-/**
- *  @author popovbodya
- */
-interface ReplenishmentView : AppView {
-
-    fun setWalletNameTitle(name: String)
-}
