+++ app/src/main/java/ru/popov/bodya/howmoney/presentation/mvp/settings/SettingsView.kt
@@ -2,4 +2,6 @@ package ru.popov.bodya.howmoney.presentation.mvp.settings
 
 import ru.popov.bodya.core.mvp.AppView
 
-interface SettingsView : AppView
+interface SettingsView : AppView {
+    fun showFavCurrency(currencyKey: String)
+}
